Refactor GPA and Attendance Sync Logic
- Updated SemesterList component styles for improved UI consistency.
- Enhanced useGpaCalculator hook to manage subject deletion confirmation.
- Modified MarksSubjectCard to improve subject name display logic and layout.
- Implemented sorting for semesters in GpaDataProvider to ensure correct order.
- Added fetchAttendanceOnly function to streamline attendance data fetching.
- Consolidated sync logic in startGradesSync and startAttendanceSync for efficiency.
- Removed deprecated grade calculation logic and related types.
- Updated UMS API interaction to handle session validation more robustly.
- Cleaned up unused parser files for attendance and dashboard.
- Adjusted constants to remove unnecessary URLs and improve clarity.

+++ ums-extension/src/lib/ums-api.ts
@@ -66,7 +66,7 @@ function parseAttendanceHtml(html: string): UMSAttendanceSummary[] {
       Slot: parseInt(tds[2].textContent?.trim() ?? '0') || 0,
       TotalDuty: parseInt(tds[3].textContent?.trim() ?? '0') || 0,
       Present: parseInt(tds[4].textContent?.trim() ?? '0') || 0,
-      Percentage: parseInt(tds[5].textContent?.trim() ?? '0') || 0,
+      Percentage: parseFloat(tds[5].textContent?.trim() ?? '0') || 0,
     });
   });
   return results;
